Migrate all account related code to the compte app.
