+++ compte/views.py
@@ -2,6 +2,7 @@ from django.contrib import messages
 from django.contrib.admin.models import CHANGE, LogEntry
 from django.contrib.auth.decorators import login_required
 from django.contrib.contenttypes.models import ContentType
+from django.core.paginator import Paginator
 from django.shortcuts import render, redirect
 from django.views.generic import TemplateView
 from django_registration.backends.activation.views import (
@@ -11,6 +12,9 @@ from django_registration.backends.activation.views import (
 
 from compte import forms
 from compte.service import create_token, validate_from_token, send_activation_mail
+from erp import versioning
+from erp.models import Erp
+from subscription.models import ErpSubscription
 
 
 class CustomActivationCompleteView(TemplateView):
@@ -96,12 +100,7 @@ def mon_email(request):
                 action_flag=CHANGE,
                 change_message=f"Demande de changement d'email {user.email} -> {new_email}",
             )
-            messages.add_message(
-                request,
-                messages.INFO,
-                f"Un email de validation vous a été envoyé à {user.email}. Merci de consulter votre boite de réception",
-            )
-            return redirect("mon_email")
+            return redirect("mon_email_sent")
     else:
         form = forms.EmailChangeForm(initial={"email": request.user.email})
     return render(
@@ -148,3 +147,75 @@ def change_email(request, activation_token):
             "compte/email_change_activation_success.html",
             context={},
         )
+
+
+@login_required
+def mes_erps(request):
+    qs = Erp.objects.select_related("accessibilite", "activite", "commune_ext").filter(
+        user_id=request.user.pk
+    )
+    published_qs = qs.published()
+    non_published_qs = qs.not_published()
+    erp_total_count = qs.count()
+    erp_published_count = published_qs.count()
+    erp_non_published_count = non_published_qs.count()
+    published = request.GET.get("published")
+    if published == "1":
+        qs = published_qs
+    elif published == "0":
+        qs = non_published_qs
+    qs = qs.filter(user_id=request.user.pk).order_by("-updated_at")
+    paginator = Paginator(qs, 10)
+    pager = paginator.get_page(request.GET.get("page", 1))
+    return render(
+        request,
+        "compte/mes_erps.html",
+        context={
+            "erp_total_count": erp_total_count,
+            "erp_published_count": erp_published_count,
+            "erp_non_published_count": erp_non_published_count,
+            "pager": pager,
+            "pager_base_url": f"?published={published or ''}",
+            "filter_published": published,
+        },
+    )
+
+
+def _mes_contributions_view(request, qs, recues=False):
+    paginator = Paginator(qs, 10)
+    pager = paginator.get_page(request.GET.get("page", 1))
+    return render(
+        request,
+        "compte/mes_contributions.html",
+        context={"pager": pager, "recues": recues},
+    )
+
+
+@login_required
+def mes_contributions(request):
+    qs = versioning.get_user_contributions(request.user)
+    return _mes_contributions_view(request, qs)
+
+
+@login_required
+def mes_contributions_recues(request):
+    qs = versioning.get_user_contributions_recues(request.user)
+    return _mes_contributions_view(request, qs, recues=True)
+
+
+@login_required
+def mes_abonnements(request):
+    qs = (
+        ErpSubscription.objects.select_related(
+            "erp", "erp__activite", "erp__commune_ext", "erp__user"
+        )
+        .filter(user=request.user)
+        .order_by("-updated_at")
+    )
+    paginator = Paginator(qs, 10)
+    pager = paginator.get_page(request.GET.get("page", 1))
+    return render(
+        request,
+        "compte/mes_abonnements.html",
+        context={"pager": pager, "pager_base_url": "?1"},
+    )
