+++ erp/templatetags/a4a.py
@@ -26,12 +26,12 @@ def active_compte_section(path, test):
     # So. Don't forget to update this code whenever we update these urls in erp.urls.
     active = any(
         [
-            test == "mon_compte" and path == "/mon_compte/",
-            test == "mes_erps" and path == "/mon_compte/erps/",
-            test == "mon_identifiant" and path == "/mon_compte/identifiant/",
-            test == "mes_contributions"
-            and path.startswith("/mon_compte/contributions/"),
-            test == "mes_abonnements" and path.startswith("/mon_compte/abonnements/"),
+            test == "mon_compte" and path == "/compte/",
+            test == "mes_erps" and path == "/compte/erps/",
+            test == "mon_identifiant" and path == "/compte/identifiant/",
+            test == "mon_email" and path.startswith("/compte/email/"),
+            test == "mes_contributions" and path.startswith("/compte/contributions/"),
+            test == "mes_abonnements" and path.startswith("/compte/abonnements/"),
             test == "mot_de_passe"
             and path
             in [
