+++ templates/compte/mon_email_sent.html
@@ -1,0 +1,12 @@
+{% extends "compte/base.html" %}
+
+{% block page_title %}Email d'activation envoyé{% endblock%}
+{% block robots %}noindex, nofollow{% endblock %}
+
+{% block compte_content %}
+  <h3 class="mb-3">Email d'activation envoyé</h3>
+  <div class="alert alert-success">
+    Un email de confirmation contenant un lien d'activation vous a été envoyé à la nouvelle
+    adresse. N'oubliez pas de cliquer ce lien pour valider et activer le changement.
+  </div>
+{% endblock %}
